requirements.txt
black==21.6b0
celery==5.2.3
Django==3.2.8
django-cors-headers==3.10.1
django-extensions==3.1.5
django-filter==21.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
django-redis==5.2.0
gunicorn==20.0.4
isort==4.3.21
markdown==3.3.6
mypy==0.910
psycopg2-binary==2.9.1
pytest==6.2.5
django-countries==6.1.3
django-simple-history==2.11.0
shortuuid==1.0.1
django-phonenumber-field==5.0.0
phonenumbers==8.12.15
django_celery_beat==2.2.1
django_celery_results==2.2.0
PyJWT==2.3.0
django-rest-knox==4.0.1
mypy==0.910
django-stubs==1.8.0
drf-yasg==1.20.0
flower==1.0.0